requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
environs==11.0.0
inflection==0.5.1
marshmallow==3.21.2
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
